fix(app): add a controller action request playback update event update
Actions like creating presets, renaming presets and adjusting sound
parameters mutated global state of the app. As a result, besides controlling
UI component, all other components would have been using the stale state.

With this action, clients can trigger an update event from PlayerManager
that will cause a state update to be propagated to all components, e.g.
if a preset is playing and it is renamed, it's updated name will be also
reflected on the Notification if the UI triggers an update event.