refactor(app): migrate HLS streams to progressive MP3 streams
ExoPlayer wasn't supporting the gapless transition from one
HLS stream to another. It is a crucial feature for the v2
sound engine. To workaround/fix it, for now, we'll be using
plain individual MP3 streams for each supported bitrate.

This effectively removes support for adaptive playback from
Noice, i.e., Noice won't be able to switch bitrates depending
on the network conditions.